chatbot/requirements.txt
# Dependencies
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.117.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.13.1
vertexai==0.0.1
Pillow==10.1.0
google-auth-oauthlib==1.2.0
requests==2.31.0


# Optional dependencies for local development
ipython==8.17.2
pytest==7.4.3